Backend/requirements.txt
Django==4.0
django-cors-headers
psycopg2-binary
djangorestframework
django-environ
PyJWT==2.6
djangorestframework-simplejwt
django-filter
Pillow
numpy
scikit-surprise
scikit-learn
pandas